pyproject.toml
[project]
name = "langchain-runnables2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.25",
    "python-dotenv>=1.1.1",
]
